requirements_main.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==1.3.1
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
arxiv==2.2.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.14.2
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
datasets==4.1.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
dotenv==0.9.9
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
fastuuid==0.13.5
feedparser==6.0.12
ffmpy==0.6.2
filelock==3.19.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
fuzz==0.1.1
google_search_results==2.4.2
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==0.3.76
langsmith==0.4.31
lazy_loader==0.4
Levenshtein==0.27.1
lightning==2.5.5
lightning-utilities==0.15.2
linkify-it-py==2.0.3
litellm==1.76.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdit-py-plugins==0.5.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
munch==4.0.0
networkx==3.4.2
nltk==3.9.1
nougat-ocr @ git+https://github.com/facebookresearch/nougat.git@5a92920d342fb6acf05fc9b594ccb4053dbe8e7a
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.109.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openreview-py==1.52.4
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.5.0
playwright==1.55.0
playwright-stealth==2.0.0
prompt_toolkit==3.0.52
propcache==0.3.2
ptyprocess==0.7.0
pyarrow==21.0.0
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
PyMuPDF==1.26.4
pypdf==6.1.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.9.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
python-socketio==5.11.2
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.3
qudida==0.0.4
questionary==2.1.1
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.14.0
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sconf==0.2.5
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.39.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
stringzilla==4.0.14
sympy==1.14.0
tenacity==9.1.2
textual==6.2.1
the==0.1.5
thefuzz==0.22.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.11.0
timm==0.5.4
tld==0.13.1
tokenizers==0.15.2
tomlkit==0.13.3
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.1
transformers==4.38.2
triton==3.4.0
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==2.3.8
wrapt==1.17.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
zstandard==0.25.0
